Make shell prompt prettier
Wrap the shell prompt across two lines with the user, host, PWD and git
status info on the first line and the command prompt underneath.

Add some pleasing colours and a unicode arrow character for the prompt.
The git shell completions will also display colours based on the git
status, but only when called from PROMPT_COMMAND, so this is now used
instead of PS1. This also requires the GIT_PS1_SHOWCOLORHINTS env var to
be set.