Set standard-resolver dirmngr option
Set the standard-resolver dirmngr option to fix "error searching
keyserver: No keyserver available" errors when searching for gpg keys on
a gpg keyserver. gpg seems not to be able to resolve DNS of the
keyserver without it.

Add dirmnmgr.conf file with the setting. Also include some useful
options for turning on debuging in gpg, which were used to debug this
problem.